프로필 기능 추가 친구 정보 요청시 재화를 long으로 변경 저금통 정보 추가

+++ src/main/java/egovframework/example/sample/web/FriendModel.java
@@ -52,14 +52,14 @@ public class FriendModel {
         return ed.get("muserid").toString();
     }
 
-    public int GetUserBalance(int idx)
+    public long GetUserBalance(int idx)
     {
         EgovMap searchin = new EgovMap();
         searchin.put("midx", idx);
 
         EgovMap ed = (EgovMap) SocketHandler.sk.sampleDAO.select("GetUserBalance", searchin);
 
-        return (int)ed.get("amount");
+        return (long)ed.get("amount");
     }
 
     public boolean Delete(String uid)
