+++ src/main/java/egovframework/example/sample/web/RoomInfo.java
@@ -3,8 +3,8 @@ package egovframework.example.sample.web;
 public class RoomInfo {
 
     public int roomnumber;
-	public int ante;
-	public int maxbet;
+	public long ante;
+	public long maxbet;
 	public int maxusersize;
     public int title = 0;
     public int currentcount;
